ART: Fix infinite recursion for deopt at dex pc 0
Previously, the interpreter checked for dex pc 0 to see if the method was
just entered. If we deopt at dex pc 0, the instrumentation would emit an
erroneous MethodEnteredEvent and the JIT would have received a MethodEntered()
call. For JIT-on-first-use, the method would be compiled the same way as
before, leading to the same deopt until stack overflow.

We fix this by using a new `from_deoptimize` flag passed by the caller.

Test: 680-checker-deopt-dex-pc-0
Test: testrunner.py --host \
      --jit --runtime-option=-Xjitthreshold:0
Bug: 62611253

Change-Id: I50b88f15484aeae16e1375a1d80f6563fb9066e7
Signed-off-by: mydongistiny <jaysonedson@gmail.com>